[add] function of showing controllable list as subcommand
